Notify users of activity involving them.
Add accounts.Notification. Notifications get created when someone
comments on your profile, likes or replies to a comment you made, or
completes a challenge you created.

They're implemented with a generic foreign key, so they can be related
to any model.

Reworked comments to use the same approach, which was a surprisingly
easy refactor, except for the South data migrations, which ended up a
little ugly.

I think reports.Activity could be done the same way, simplifying or
eliminating reports.actions.ActivityLog. The activity feed would be a
little more robust if URLs weren't stored in the database, but resolved
at run time.