Fix bugs found by Eric.
-- when a user has no points and, the point scale says, 100 points left
to token, there is a little green showing as if to suggest there has
already been some progress.

-- following is not working.  i get a system error.

   This was due to him not having a first name; that case is handled
   better now.

-- when spending coins, can the "spend" button be grayed out if the user
doesn't have any tokens.  Currently, it delivers an odd bullet point on
the top of the screen that says "no coins available to spend."

   Both spend and take back are now disabled as necessary; if a POST
   should get through the error is formatted better.

--when clicking on one's personal profile, it goes right to "update
profile."  This should go to the regular outward facing profile screen
with an "edit" button.

-- there is a character limit on comments within challenges, but it is
not there in profile comments.

   Moved the character count out of the form field's help_text and into
   the comment templates.