Fix comment liking Unicode bug.
Activities with non-ASCII characters in their names (and the one that
prompted this was in English; it just had a fancy single quote instead
of the plain apostrophe) prevented the liking of comments on their
answers, because answers' __unicode__ methods were trying to interpolate
Unicode activity names into byte strings. Making answers' __unicode__
methods actually produce Unicode fixed this.